[sshfs_mounts] Better organize related files
