+++ src/CMakeLists.txt
@@ -31,7 +31,7 @@ add_subdirectory(rpc)
 add_subdirectory(settings)
 add_subdirectory(simplestreams)
 add_subdirectory(ssh)
-add_subdirectory(sshfs_mount_handler)
+add_subdirectory(sshfs_mount)
 add_subdirectory(utils)
 add_subdirectory(blueprint_provider)
 add_subdirectory(xz_decoder)
