+++ src/platform/backends/shared/base_virtual_machine_factory.h
@@ -21,7 +21,7 @@
 #include <multipass/exceptions/not_implemented_on_this_backend_exception.h>
 #include <multipass/format.h>
 #include <multipass/logging/log.h>
-#include <multipass/mount_handlers/mount_handler.h>
+#include <multipass/mount_handler.h>
 #include <multipass/virtual_machine_factory.h>
 
 #include <daemon/default_vm_image_vault.h>
