+++ tests/test_sftpserver.cpp
@@ -27,8 +27,9 @@
 #include "temp_dir.h"
 #include "temp_file.h"
 
+#include <src/sshfs_mount/sftp_server.h>
+
 #include <multipass/format.h>
-#include <multipass/mount_handlers/sshfs/sftp_server.h>
 #include <multipass/platform.h>
 #include <multipass/ssh/ssh_session.h>
